requirements.txt
dnspython==2.3.0
PyYAML==6.0
